requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.0.6
et-xmlfile==1.1.0
Flask==2.1.2
Flask-RESTful==0.3.9
fpdf==1.7.2
google-auth==2.9.0
google-auth-oauthlib==0.5.2
gspread==5.4.0
idna==3.3
image==1.5.33
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.1
oauthlib==3.2.0
openpyxl==3.0.10
pandas==1.4.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
pytz==2022.1
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.10
Werkzeug==2.1.2
zipp==3.8.0
